RPI4 clone and build scripts
This is an initial version of the build scripts I've been
using for the rpi4, they can be tweaked for the rpi3
with little changes or release/debug builds with
differing flags. I tend to have a heap of these
lying about and just run the the variation i'm testing
or location the result will be copied to/etc.

They are provided as an example for people to begin
developing the rpi4 firmware but don't replace
the pftf build or edk2 build instructions which
should be referred to for details about what
is happening here.

Signed-off-by: Jeremy Linton <lintonrjeremy@gmail.com>